requirements.txt
# Gate A foundation
pyyaml>=6.0
psutil>=5.9

# Gate B/C - Local model (Forge) & embeddings
python-dotenv>=1.0
# Forge uses llama-cpp-python. For Qwen3-VL vision: install JamePeng fork 0.3.24 (see RUN.md)
#   pip install llama-cpp-python  # standard (no vision)
#   pip install llama-cpp-python @ git+https://github.com/jamepeng/llama-cpp-python  # Qwen3-VL vision
llama-cpp-python
sentence-transformers  # Embeddings for vector store
huggingface_hub    # scripts/download_model.py

# Gate B Phase 2 - Memory (LanceDB)
lancedb
pyarrow

# Gate B Phase 3 - Grok (x.ai, OpenAI-compatible)
openai>=1.0

# Local web search (free, no API key; ddgs is the maintained successor to duckduckgo-search)
ddgs>=9.0

# Gate C - Desktop automation
pyautogui>=0.9.54
# Gate C Phase 2 - Browser automation
playwright>=1.40.0

pynput>=1.7.6
pywinauto>=0.6.8
pillow>=10.0.0

# Gate B+ (later)
# anthropic         # Claude API
